HSEARCH-3692 Move the ISBN data type to a common package in documentation tests
We're going to use this type a lot.
