+++ documentation/src/test/java/org/hibernate/search/documentation/mapper/orm/identifiermapping/customtype/Book.java
@@ -10,6 +10,8 @@ import javax.persistence.Convert;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 
+import org.hibernate.search.documentation.testsupport.data.ISBN;
+import org.hibernate.search.documentation.testsupport.data.ISBNAttributeConverter;
 import org.hibernate.search.mapper.pojo.bridge.mapping.annotation.IdentifierBridgeRef;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.DocumentId;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
