+++ documentation/src/test/java/org/hibernate/search/documentation/mapper/orm/identifiermapping/customtype/ISBNIdentifierBridge.java
@@ -6,6 +6,7 @@
  */
 package org.hibernate.search.documentation.mapper.orm.identifiermapping.customtype;
 
+import org.hibernate.search.documentation.testsupport.data.ISBN;
 import org.hibernate.search.mapper.pojo.bridge.IdentifierBridge;
 import org.hibernate.search.mapper.pojo.bridge.runtime.IdentifierBridgeFromDocumentIdentifierContext;
 import org.hibernate.search.mapper.pojo.bridge.runtime.IdentifierBridgeToDocumentIdentifierContext;
@@ -20,7 +21,7 @@ public class ISBNIdentifierBridge implements IdentifierBridge<ISBN> {
 	@Override
 	public ISBN fromDocumentIdentifier(String documentIdentifier,
 			IdentifierBridgeFromDocumentIdentifierContext context) {
-		return documentIdentifier == null ? null : new ISBN( documentIdentifier );
+		return documentIdentifier == null ? null : ISBN.parse( documentIdentifier );
 	}
 
 }
