+++ documentation/src/test/java/org/hibernate/search/documentation/testsupport/data/ISBN.java
@@ -4,14 +4,25 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.documentation.mapper.orm.identifiermapping.customtype;
+package org.hibernate.search.documentation.testsupport.data;
 
 import java.io.Serializable;
+import java.util.regex.Pattern;
 
 public class ISBN implements Serializable {
+
+	private static final Pattern ISBN_13_PATTERN = Pattern.compile( "\\d{3}-\\d-\\d{2}-\\d{6}-\\d" );
+
+	public static ISBN parse(String string) {
+		if ( string == null || !ISBN_13_PATTERN.matcher( string ).matches() ) {
+			throw new IllegalArgumentException( "Not an ISBN: " + string );
+		}
+		return new ISBN( string );
+	}
+
 	private final String stringValue;
 
-	public ISBN(String stringValue) {
+	private ISBN(String stringValue) {
 		this.stringValue = stringValue;
 	}
